bs3kit: 8086 friendly 32-bit and 64-bit unsigned integer division.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@60485 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/uint32.h
@@ -928,8 +928,8 @@ DECLINLINE(PRTUINT32U) RTUInt32BitSetRange(PRTUINT32U pValueResult, unsigned iFi
  */
 DECLINLINE(bool) RTUInt32BitAreAllSet(PRTUINT32U pValue)
 {
-    return pValue->s.Hi == UINT32_MAX
-        && pValue->s.Lo == UINT32_MAX;
+    return pValue->s.Hi == UINT16_MAX
+        && pValue->s.Lo == UINT16_MAX;
 }
